PreprocessGraph is templated for MarkerType
